Worked on 3D blast wave test problem.

+++ source/ics/icgen.cpp
@@ -389,7 +389,7 @@ int main(int argc, char **argv)
 #endif
 #ifdef PARALLEL
     cout <<outfile <<"\n";
-    dataio=0; dataio=new dataio_silo_pllel ("FLOAT",&MCMD);
+    dataio=0; dataio=new dataio_silo_pllel ("DOUBLE",&MCMD);
 #endif
   }
 #endif // if SILO defined.
